First pass of 'mlog' support
I don't love this factoring of the API yet, but this is capable of
writing a qlog file with the MoQ event extensions and logs the creation
of a SERVER_SETUP message at least.

Likely to refactor things a bit in the next commit now that we've proven
this out end to end.