Update interop relay instance config to allow TCP
Enables forwarding external tcp/443 to internal tcp/4443 so our app can
receive HTTPS traffic to the simple (TCP only) HTTPS handler.

This endpoint was originally added to simply serve fingerprints to allow
WebTransport connections to untrusted certificates, but we've recently
extended it to allow access to session logs to aid with interop testing.